+++ src/main/java/io/github/astrapi69/gen/tree/handler/IBaseTreeNodeHandlerExtensions.java
@@ -26,6 +26,7 @@ package io.github.astrapi69.gen.tree.handler;
 
 import java.util.concurrent.atomic.AtomicReference;
 
+import io.github.astrapi69.gen.tree.visitor.MergeTreeNodesVisitor;
 import lombok.NonNull;
 import io.github.astrapi69.gen.tree.api.IBaseTreeNode;
 
@@ -70,4 +71,31 @@ public class IBaseTreeNodeHandlerExtensions
 		});
 		return found.get();
 	}
+
+	/**
+	 * Merge the given tree nodes with the given root tree node
+	 *
+	 * @param <V>
+	 *            the generic type of the value
+	 * @param <K>
+	 *            the generic type of the id of the node
+	 * @param <T>
+	 *            the generic type of the concrete tree node
+	 * @param root
+	 *            the root tree node
+	 * @param treeNodes
+	 *            the tree nodes to merge with
+	 * @return the root with the merged tree nodes
+	 */
+	public static <V, K, T extends IBaseTreeNode<V, K, T>> T mergeTreeNodes(final @NonNull T root,
+		T... treeNodes)
+	{
+		MergeTreeNodesVisitor<V, K, T> visitor = new MergeTreeNodesVisitor<>(root);
+		for (T treeNode : treeNodes)
+		{
+			treeNode.accept(visitor);
+		}
+		return root;
+	}
+
 }